refactor(cli): split cli/main.py (1989 LOC) by command group (#166)
Mechanical split — no behavior changes. main.py becomes thin Typer root
(210 LOC) and all command bodies move under commands/.

- commands/jobs.py: sbatch, squeue, scancel, sinfo, tail
- commands/reports.py: sacct, sreport
- commands/config.py: config show/paths/init (sub-Typer)
- commands/templates.py: template list/show (sub-Typer)
- commands/ui.py: ui + _run_ui_dev
- _helpers/debug_callback.py: DebugCallback (re-exported from main.py
  for cli/workflow.py)
- _helpers/sbatch_helpers.py: _submit_via_transport,
  _build_extra_sbatch_args, _parse_gres_gpu,
  _print_in_place_sync_preview, _parse_env_vars, _parse_bool,
  _parse_container_args

main.py keeps flow_run shim + backward-compat re-exports for test
monkeypatching (Slurm, get_config, get_config_paths, DebugCallback,
_parse_env_vars, _parse_container_args). Command modules dereference
Slurm/get_config/get_config_paths via sys.modules["srunx.cli.main"]
at call time so the existing patch("srunx.cli.main.Slurm") fixtures
continue to intercept — the dotted-submodule import trick doesn't
work because srunx.cli.__init__ re-exports a main() function that
shadows the main submodule.

Verified:
- All `srunx --help` / per-command --help outputs byte-identical to
  pre-refactor.
- 1967 tests pass (3 pre-existing failures on main due to local
  config side-effects, unrelated).
- mypy . clean (298 files).
- ruff check + format clean.
- tests/architecture/test_import_layers.py passes.

Note: commands/jobs.py is 900 LOC (exceeds the 400 LOC guard). The
irreducible size is sbatch itself (~500 LOC of Typer annotations);
further reduction would require either splitting sbatch across
multiple command functions (behavior risk) or splitting jobs.py into
sbatch.py + queue.py (deviates from the target layout in #166).

Closes #166
